+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { StoragePanel } from "./StoragePanel";
+export { StatusBanner } from "./StatusBanner";
+export { StorageSection } from "./StorageSection";
+export { AdvancedSection } from "./AdvancedSection";
+export { DetailsSection } from "./DetailsSection";
+export { StatusCard } from "./StatusCard";